+++ src/components/index.ts
@@ -2,6 +2,5 @@ export { default as ErrorMessage } from './errorMessage';
 export { default as DetailCard } from './inspectModules';
 export { default as WhatsThisLogo } from './logo';
 export { DrawMenu, HomeMenu } from './menu';
-export { IPAddressGeo, MyIpAddressModal, UserAgentModel } from './modals';
 export { GeneralTemplate, ReportDNSError, ReportInspectionError } from './reportButton';
 export { ServiceIcon } from './ServiceIcon';
