+++ src/modals/IPAddress.tsx
@@ -1,115 +1,8 @@
-import {
-	Typography, Button, Grid, Link, IconButton, Stack, styled,
-	Dialog, DialogTitle, DialogContent
-} from "@mui/material";
 import { useContext, useEffect, useState } from "react";
-import { DialogTitleProps } from "../interfaces";
 import { ConnectionContext } from "../context";
-import { IPAddresses, IPGeolocation, getBothIPAddresses, getIPGeolocation, getUserAgent } from "libwhatsthis";
-
-import CloseIcon from '@mui/icons-material/Close';
-
-const BootstrapDialog = styled(Dialog)(({ theme }) => ({
-	'& .MuiDialogContent-root': {
-		padding: theme.spacing(2),
-	},
-	'& .MuiDialogActions-root': {
-		padding: theme.spacing(1),
-	},
-}));
-
-const BootstrapDialogTitle = (props: DialogTitleProps) => {
-	const { children, onClose, ...other } = props;
-
-	return (
-		<DialogTitle sx={{ m: 0, p: 2 }} {...other}>
-			{children}
-			{onClose && (
-				<IconButton
-					aria-label="close"
-					onClick={onClose}
-					sx={{
-						position: 'absolute',
-						right: 8,
-						top: 8,
-						color: (theme) => theme.palette.grey[500],
-					}}
-				>
-					<CloseIcon />
-				</IconButton>
-			)}
-		</DialogTitle>
-	);
-}
-
-export const UserAgentModel = () => {
-	const [open, setOpen] = useState(false);
-	const handleOpen = () => setOpen(true);
-	const handleClose = () => setOpen(false);
-
-	const userAgent = getUserAgent(window.navigator.userAgent);
-
-	return (
-		<div>
-			<Button onClick={handleOpen} variant="outlined">Connection Info</Button>
-			<BootstrapDialog
-				open={open}
-				onClose={handleClose}
-				aria-labelledby="conn-modal-modal-title"
-				aria-describedby="conn-modal-modal-description"
-			>
-				<BootstrapDialogTitle id="conn-modal-modal-title" onClose={handleClose}>
-					UserAgent Information
-				</BootstrapDialogTitle>
-				<DialogContent>
-					<Grid container id="conn-modal-modal-description" spacing={2} my={2}>
-						<Grid item xs={12} sm={4}>
-							<Typography fontWeight={700}>Your UserAgent</Typography>
-						</Grid>
-						<Grid item xs={12} sm={8}>
-							<Typography>{window.navigator.userAgent}</Typography>
-						</Grid>
-						<Grid item xs={12} sm={12}>
-							<Typography>
-								From this, it can be determined:
-							</Typography>
-						</Grid>
-						<Grid item xs={12} sm={4}>
-							<Typography fontWeight={700}>Browser</Typography>
-						</Grid>
-						<Grid item xs={12} sm={8}>
-							<Typography>{userAgent.browser.name} {userAgent.browser.version}</Typography>
-						</Grid>
-						<Grid item xs={12} sm={4}>
-							<Typography fontWeight={700}>Engine</Typography>
-						</Grid>
-						<Grid item xs={12} sm={8}>
-							<Typography>{userAgent.engine.name} {userAgent.engine.version}</Typography>
-						</Grid>
-						<Grid item xs={12} sm={4}>
-							<Typography fontWeight={700}>Operating System</Typography>
-						</Grid>
-						<Grid item xs={12} sm={8}>
-							<Typography>{userAgent.system.name} {userAgent.system.version}</Typography>
-						</Grid>
-						<Grid item xs={12} sm={4}>
-							<Typography fontWeight={700}>Device</Typography>
-						</Grid>
-						<Grid item xs={12} sm={8}>
-							<Typography>{userAgent.device ?? <em>Unspecified</em>}</Typography>
-						</Grid>
-						<Grid item xs={12} sm={4}>
-							<Typography fontWeight={700}>CPU</Typography>
-						</Grid>
-						<Grid item xs={12} sm={8}>
-							<Typography>{userAgent.cpu ?? <em>Unspecified</em>}</Typography>
-						</Grid>
-					</Grid>
-				</DialogContent>
-			</BootstrapDialog>
-		</div>
-	);
-}
+import { IPAddresses, IPGeolocation, getBothIPAddresses, getIPGeolocation } from "libwhatsthis";
+import { Button, DialogContent, Grid, IconButton, Link, Stack, Typography } from "@mui/material";
+import { BootstrapDialog, BootstrapDialogTitle } from "./_shared";
 
 export const MyIpAddressModal = () => {
 	const { connectionState } = useContext(ConnectionContext);
