+++ src/modals/_shared.tsx
@@ -1,0 +1,36 @@
+import { Dialog, DialogTitle, IconButton, styled } from "@mui/material";
+import { DialogTitleProps } from "../interfaces";
+import CloseIcon from '@mui/icons-material/Close';
+
+export const BootstrapDialog = styled(Dialog)(({ theme }) => ({
+	'& .MuiDialogContent-root': {
+		padding: theme.spacing(2),
+	},
+	'& .MuiDialogActions-root': {
+		padding: theme.spacing(1),
+	},
+}));
+
+export const BootstrapDialogTitle = (props: DialogTitleProps) => {
+	const { children, onClose, ...other } = props;
+
+	return (
+		<DialogTitle sx={{ m: 0, p: 2 }} {...other}>
+			{children}
+			{onClose && (
+				<IconButton
+					aria-label="close"
+					onClick={onClose}
+					sx={{
+						position: 'absolute',
+						right: 8,
+						top: 8,
+						color: (theme) => theme.palette.grey[500],
+					}}
+				>
+					<CloseIcon />
+				</IconButton>
+			)}
+		</DialogTitle>
+	);
+}
