+++ src/modals/index.ts
@@ -1,0 +1,2 @@
+export { default as UserAgentModel } from './UserAgent';
+export { IPAddressGeo, MyIpAddressModal } from './IPAddress';
